# docs/reference/kumod.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "kumod",
    "description": "",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "2023.11.18-8b8f5312"
  },
  "paths": {
    "/api/admin/bounce/v1": {
      "get": {
        "tags": [
          "bounce"
        ],
        "summary": "Allows the system operator to list all currently active administrative bounces that have been",
        "description": "Allows the system operator to list all currently active administrative bounces that have been\nconfigured.",
        "operationId": "bounce_v1_list",
        "responses": {
          "200": {
            "description": "Returned information about current admin bounces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BounceV1ListEntry"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "bounce"
        ],
        "summary": "Allows the system operator to administratively bounce messages that match",
        "description": "Allows the system operator to administratively bounce messages that match\ncertain criteria, or if no criteria are provided, ALL messages.",
        "operationId": "bounce_v1",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BounceV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bounce added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BounceV1Response"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "bounce"
        ],
        "summary": "Allows the system operator to delete an administrative bounce entry by its id.",
        "description": "Allows the system operator to delete an administrative bounce entry by its id.",
        "operationId": "bounce_v1_delete",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BounceV1CancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Removed the requested bounce id"
          },
          "404": {
            "description": "The requested bounce id is no longer, or never was, valid"
          }
        }
      }
    },
...
    "/api/admin/set_diagnostic_log_filter/v1": {
      "post": {
        "tags": [
          "logging"
        ],
        "summary": "Changes the diagnostic log filter dynamically.",
        "description": "Changes the diagnostic log filter dynamically.\nSee <https://docs.kumomta.com/reference/kumo/set_diagnostic_log_filter/>\nfor more information on diagnostic log filters.",
        "operationId": "set_diagnostic_log_filter_v1",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDiagnosticFilterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Diagnostic level set successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BounceV1CancelRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "BounceV1ListEntry": {
        "type": "object",
        "required": [
          "id",
          "reason",
          "duration",
          "bounced",
          "total_bounced"
        ],
        "properties": {
          "bounced": {
            "type": "object",
            "description": "A map of queue name to number of bounced messages that\nwere processed by this entry since it was created.",
            "additionalProperties": {
              "type": "integer",
              "minimum": 0
            },
            "example": {
              "gmail.com": 200,
              "yahoo.com": 100
            }
          },
          "campaign": {
            "type": "string",
            "description": "The campaign field of the original request, if any.",
            "nullable": true
          },
          "domain": {
            "type": "string",
            "description": "The domain field of the original request, if any.",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "description": "The time remaining until this entry expires and is automatically\nremoved."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The id of this bounce rule. Corresponds to the `id` field\nreturned by the originating request that set up the bounce,\nand can be used to identify this particular entry if you\nwish to delete it later.",
            "example": "552016f1-08e7-4e90-9da3-fd5c25acd069"
          },
          "reason": {
            "type": "string",
            "description": "The reason field of the original request"
          },
          "routing_domain": {
            "type": "string",
            "description": "The routing_domain field of the original request, if any.",
            "nullable": true
          },
          "tenant": {
            "type": "string",
            "description": "The tenant field of the original request, if any.",
            "nullable": true
          },
          "total_bounced": {
            "type": "integer",
            "description": "The sum of the number of bounced messages reported by\nthe `bounced` field.",
            "minimum": 0
          }
        }
      },
      "BounceV1Request": {
        "type": "object",
        "description": "Describes which messages should be bounced.\nThe criteria apply to the scheduled queue associated\nwith a given message.",
        "required": [
          "reason"
        ],
        "properties": {
          "campaign": {
            "type": "string",
            "description": "The campaign name to match. If omitted, any campaign will match.",
            "nullable": true
          },
          "domain": {
            "type": "string",
            "description": "The domain name to match. If omitted, any domain will match.",
            "example": "example.com",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "description": "Defaults to \"5m\". Specifies how long this bounce directive remains active.\nWhile active, newly injected messages that match the bounce criteria\nwill also be bounced.",
            "example": "20m",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "description": "Reason to log in the delivery log",
            "example": "Cleaning up a bad send"
          },
          "routing_domain": {
            "type": "string",
            "description": "The routing_domain name to match. If omitted, any routing_domain will match.",
            "nullable": true
          },
          "tenant": {
            "type": "string",
            "description": "The tenant to match. If omitted, any tenant will match.",
            "nullable": true
          }
        }
      },
      "BounceV1Response": {
        "type": "object",
        "required": [
          "id",
          "bounced",
          "total_bounced"
        ],
        "properties": {
          "bounced": {
            "type": "object",
            "description": "A map of queue name to number of bounced messages that\nwere processed as part of the initial sweep.\nAdditional bounces may be generated if/when other messages\nthat match the rule are discovered, but those obviously\ncannot be reported in the context of the initial request.",
            "additionalProperties": {
              "type": "integer",
              "minimum": 0
            },
            "example": {
              "gmail.com": 200,
              "yahoo.com": 100
            }
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The id of the bounce rule that was registered.\nThis can be used later to delete the rule if desired.",
            "example": "552016f1-08e7-4e90-9da3-fd5c25acd069"
          },
          "total_bounced": {
            "type": "integer",
            "description": "The sum of the number of bounced messages reported by\nthe `bounced` field.",
            "example": 300,
            "minimum": 0
          }
        }
      },
...
      "SetDiagnosticFilterRequest": {
        "type": "object",
        "required": [
          "filter"
        ],
        "properties": {
          "filter": {
            "type": "string",
            "description": "The diagnostic filter spec to use",
            "example": "kumod=trace"
          }
        }
      },
...
    },
...
  }
}